separated image uploader into component
